LPD-8519 Remove service from SharepointRepositoryTokenBrokerFactory, rename it to SharepointRepositoryTokenBrokerFactoryUtil, pass configurationAdmin as a parameter and apply to usages
