+++ modules/dxp/apps/sharepoint-rest/sharepoint-rest-repository/src/main/java/com/liferay/sharepoint/rest/repository/internal/document/library/repository/authorization/oauth2/util/SharepointRepositoryTokenBrokerFactoryUtil.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.sharepoint.rest.repository.internal.document.library.repository.authorization.oauth2;
+package com.liferay.sharepoint.rest.repository.internal.document.library.repository.authorization.oauth2.util;
 
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.sharepoint.rest.repository.internal.configuration.SharepointRepositoryConfiguration;
+import com.liferay.sharepoint.rest.repository.internal.document.library.repository.authorization.oauth2.SharepointRepositoryTokenBroker;
 
 import java.io.IOException;
 
@@ -17,32 +18,34 @@ import java.util.NoSuchElementException;
 import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.service.cm.Configuration;
 import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Adolfo Pérez
  */
-@Component(service = SharepointRepositoryTokenBrokerFactory.class)
-public class SharepointRepositoryTokenBrokerFactory {
+public class SharepointRepositoryTokenBrokerFactoryUtil {
 
-	public SharepointRepositoryTokenBroker create(
+	public static SharepointRepositoryTokenBroker create(
+		ConfigurationAdmin configurationAdmin, String configurationPid) {
+
+		return create(
+			_getSharepointRepositoryConfiguration(
+				configurationAdmin, configurationPid));
+	}
+
+	public static SharepointRepositoryTokenBroker create(
 		SharepointRepositoryConfiguration sharepointRepositoryConfiguration) {
 
 		return new SharepointRepositoryTokenBroker(
 			sharepointRepositoryConfiguration);
 	}
 
-	public SharepointRepositoryTokenBroker create(String configurationPid) {
-		return create(_getSharepointRepositoryConfiguration(configurationPid));
-	}
-
-	private SharepointRepositoryConfiguration
-		_getSharepointRepositoryConfiguration(String configurationPid) {
+	private static SharepointRepositoryConfiguration
+		_getSharepointRepositoryConfiguration(
+			ConfigurationAdmin configurationAdmin, String configurationPid) {
 
 		try {
 			Configuration[] configurations =
-				_configurationAdmin.listConfigurations(
+				configurationAdmin.listConfigurations(
 					"(service.factoryPid=" +
 						SharepointRepositoryConfiguration.class.getName() +
 							")");
@@ -66,8 +69,5 @@ public class SharepointRepositoryTokenBrokerFactory {
 			throw new SystemException(exception);
 		}
 	}
-
-	@Reference
-	private ConfigurationAdmin _configurationAdmin;
 
 }
